rename BlogList to PostsList
