+++ tsconfig.json
@@ -19,7 +19,10 @@
     ],
     "paths": {
       "@/*": ["./src/*"],
-      "@public/*": ["./public/*"]
+      "@public/*": ["./public/*"],
+      "@components/*": ["./src/app/ui/components/*"],
+      "@contexts/*": ["./src/app/ui/contexts/*"],
+      "@features/*": ["./src/app/ui/features/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
